+++ deploy/demo/kind-with-registry.sh
@@ -16,6 +16,8 @@ cat <<EOF | kind create cluster --config=-
 kind: Cluster
 apiVersion: kind.x-k8s.io/v1alpha4
 name: kobs-demo
+featureGates:
+  EphemeralContainers: true
 containerdConfigPatches:
 - |-
   [plugins."io.containerd.grpc.v1.cri".registry.mirrors."localhost:${reg_port}"]