+++ plugins/kiali/pkg/instance/instance.go
@@ -22,20 +22,20 @@ var (
 
 // Config is the structure of the configuration for a single Kiali instance.
 type Config struct {
-	Name        string  `yaml:"name"`
-	DisplayName string  `yaml:"displayName"`
-	Description string  `yaml:"description"`
-	Address     string  `yaml:"address"`
-	Username    string  `yaml:"username"`
-	Password    string  `yaml:"password"`
-	Token       string  `yaml:"token"`
-	Traffic     Traffic `yaml:"traffic"`
+	Name        string  `json:"name"`
+	DisplayName string  `json:"displayName"`
+	Description string  `json:"description"`
+	Address     string  `json:"address"`
+	Username    string  `json:"username"`
+	Password    string  `json:"password"`
+	Token       string  `json:"token"`
+	Traffic     Traffic `json:"traffic"`
 }
 
 // Traffic is the traffic configuration, to set the value to mark a edge as degraded and failure.
 type Traffic struct {
-	Degraded float64 `yaml:"degraded"`
-	Failure  float64 `yaml:"failure"`
+	Degraded float64 `json:"degraded"`
+	Failure  float64 `json:"failure"`
 }
 
 // Graph is the structure of the returned topology graph from Kiali. We have to implement it by ourselve, because we